pyproject.toml
[tool.poetry]
name = "ingredex"
version = "0.1.0"
description = "Ingredex backend API"
authors = []
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-dotenv = "^1.0.0"
pydantic-settings = "^2.3.0"
pydantic = { extras = ["email"], version = "^2.7.0" }
httpx = "^0.27.0"
redis = "^5.0.6"
python-multipart = "^0.0.9"
loguru = "^0.7.2"
tenacity = "^8.4.1"
setuptools = ">=69.0.0"
crewai = ">=0.80.0"
langchain-groq = ">=0.2.0"
langchain = ">=0.3.0"
langchain-core = ">=0.3.0"
litellm = "^1.82.6"

[tool.poetry.scripts]
dev = "app.cli:dev"
migrate = "app.cli:migrate"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
